Make the plugin configuration-cache-friendly (#7)
* Make the plugin configuration-cache-friendly

Signed-off-by: Octol1ttle <l1ttleofficial@outlook.com>

* make "classes" depend on yamlang instead of "jar"
the reason this is needed is because our task writes to build/resources/ and that is also used by other tasks. Gradle doesn't like that and forces us to make one of the later tasks depend on our task. "jar" was used during initial testing but that doesn't work when executing an IDE run

Signed-off-by: Octol1ttle <l1ttleofficial@outlook.com>

* convert spaces to tabs

Signed-off-by: Octol1ttle <l1ttleofficial@outlook.com>

* unbump version

Signed-off-by: Octol1ttle <l1ttleofficial@outlook.com>

---------

Signed-off-by: Octol1ttle <l1ttleofficial@outlook.com>